requirements.txt
Pillow==8.4.0
pytesseract==0.3.8
python-dotenv
moviepy